fix db ~02

+++ Bodies/Req/Specific/ReqBody_BangDiemHocPhan.cs
@@ -2,13 +2,13 @@
 {
     public record class ReqBody_BangDiemHocPhan : BaseReqBody<BangDiemHocPhan>
     {
-        public long ? MaBangDiemHocPhan { get; set; }
-        public long ? MaHocPhan         { get; set; }
-        public long ? MaSinhVien        { get; set; }
-        public float? DiemQuaTrinh      { get; set; }
-        public float? DiemGiuaKy        { get; set; }
-        public float? DiemThucHanh      { get; set; }
-        public float? DiemCuoiKy        { get; set; }
+        public long   ? MaBangDiemHocPhan { get; set; }
+        public long   ? MaHocPhan         { get; set; }
+        public long   ? MaSinhVien        { get; set; }
+        public decimal? DiemQuaTrinh      { get; set; }
+        public decimal? DiemGiuaKy        { get; set; }
+        public decimal? DiemThucHanh      { get; set; }
+        public decimal? DiemCuoiKy        { get; set; }
 
         public override Expression<Func<BangDiemHocPhan, bool>> MatchExpression()
         {
