+++ Bodies/Req/Specific/ReqBody_KetQuaHocTap.cs
@@ -2,11 +2,11 @@
 {
     public record class ReqBody_KetQuaHocTap : BaseReqBody<KetQuaHocTap>
     {
-        public long  ? MaKetQuaHocTap     { get; set; }
-        public float ? DiemTrungBinhHocKy { get; set; }
-        public string? XepLoaiHocTap      { get; set; }
-        public long  ? MaHocKyNamHoc      { get; set; }
-        public long  ? MaSinhVien         { get; set; }
+        public long   ? MaKetQuaHocTap     { get; set; }
+        public decimal? DiemTrungBinhHocKy { get; set; }
+        public string ? XepLoaiHocTap      { get; set; }
+        public long   ? MaHocKyNamHoc      { get; set; }
+        public long   ? MaSinhVien         { get; set; }
 
         public override Expression<Func<KetQuaHocTap, bool>> MatchExpression()
         {
